refactor: menu, menu-group 병합
